refactor: 기사님 상세 webp 변경
