+++ src/components/searchMover/MoverCard.tsx
@@ -7,8 +7,8 @@ import Image from "next/image";
 import { MoveTypeLabel } from "../common/chips/MoveTypeLabel";
 import { useWindowWidth } from "@/hooks/useWindowWidth";
 import { IMoverInfo } from "@/types/mover.types";
-import defaultProfileLg from "@/assets/img/mascot/profile-lg.png";
-import defaultProfileSm from "@/assets/img/mascot/profile-sm.png";
+import defaultProfileLg from "@/assets/img/mascot/moverprofile-lg.webp";
+import defaultProfileSm from "@/assets/img/mascot/moverprofile-sm.webp";
 import badge from "@/assets/icon/etc/icon-chat.png";
 import star from "@/assets/icon/star/icon-star-active-lg.png";
 import like from "@/assets/icon/like/icon-like-red.png";
